#8 finished score

+++ projectBoombastic/Assets/Scripts/CollectCoin.cs
@@ -6,7 +6,7 @@ public class CollectCoin : MonoBehaviour
 {
 
     [SerializeField] private List<GameObject> coinPrefabs;
-    public float coin = 0;
+    public int coin = 0;
     private Vector2 screenBounds;
 
     private GameObject coinPrefab;
